backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
redis
pydantic
python-jose
passlib[bcrypt]
python-dotenv 
python-multipart
PyMuPDF
requests